+++ android/app/src/main/java/com/on/turip/ui/trip/detail/TripDetailActivity.kt
@@ -15,13 +15,13 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
 import com.on.turip.R
 import com.on.turip.databinding.ActivityTripDetailBinding
+import com.on.turip.ui.common.TuripUrlConverter
 import com.on.turip.ui.common.base.BaseActivity
 import com.on.turip.ui.common.loadCircularImage
 import com.on.turip.ui.trip.detail.webview.TuripWebChromeClient
 import com.on.turip.ui.trip.detail.webview.TuripWebViewClient
 import com.on.turip.ui.trip.detail.webview.WebViewVideoBridge
 import com.on.turip.ui.trip.detail.webview.applyVideoSettings
-import com.on.turip.ui.trip.detail.webview.extractVideoId
 
 class TripDetailActivity : BaseActivity<TripDetailViewModel, ActivityTripDetailBinding>() {
     override val binding: ActivityTripDetailBinding by lazy {
@@ -196,7 +196,7 @@ class TripDetailActivity : BaseActivity<TripDetailViewModel, ActivityTripDetailB
             binding.wvTripDetailVideo.apply {
                 addJavascriptInterface(
                     WebViewVideoBridge(
-                        url.extractVideoId(),
+                        TuripUrlConverter.extractVideoId(url),
                     ) { showWebViewErrorView() },
                     BRIDGE_NAME_IN_JS_FILE,
                 )
